refactor!: rename vertices to edges for clarity (#60)
`vertices` can be confused by the corners of a square in the square
lattice, sharing some mind space with the concept of site. `edges` on
the other hand are more common in graph lingo to refer to the
connections between two sites. Here we're renaming the concept of vertex
to edge for clarity.